refactor: extract flowrs-airflow as workspace crate
