+++ Cargo.toml
@@ -46,20 +46,16 @@ ansi-to-tui = { version = "8.0.1" }
 anyhow = { workspace = true }
 flowrs-config = { path = "crates/flowrs-config" }
 flowrs-airflow-model = { path = "crates/flowrs-airflow-model" }
+flowrs-airflow = { path = "crates/flowrs-airflow" }
 async-trait = { workspace = true }
-aws-config = { version = "1.8.11" }
-aws-sdk-mwaa = "1.99.0"
 chrono = { workspace = true }
 clap = { workspace = true }
 crossterm = "0.29.0"
 dirs = { workspace = true }
 futures = "0.3.32"
-google-cloud-auth = { version = "1.6.0", default-features = false, features = ["default-rustls-provider"] }
 inquire = "0.9.3"
 log = { workspace = true }
 ratatui = { version = "0.30.0", features = ["unstable-widget-ref"] }
-regex = "1.12.3"
-reqwest = { workspace = true }
 serde = { workspace = true }
 serde_json = { workspace = true }
 strum = { workspace = true }
@@ -74,6 +70,7 @@ webbrowser = "1.2.0"
 
 [dev-dependencies]
 mockito = "1.7.1"
+reqwest = { workspace = true }
 
 [profile.release]
 opt-level = "z"
