+++ crates/flowrs-airflow/src/client/auth/static_token.rs
